confirmed, RDCSS is necessary.
to keep invariant from root, root and its left child must be updated. otherwise, when iteration reaches the left child of root and decide to change its left child to new, the gcas operation will be bound to fail. because the gen of the left child of root is old.